+++ lib/health_cards.rb
@@ -6,6 +6,8 @@ require 'health_cards/issuer'
 require 'health_cards/key'
 require 'health_cards/jws'
 require 'health_cards/key_set'
+require 'health_cards/private_key'
+require 'health_cards/public_key'
 require 'health_cards/health_card'
 require 'health_cards/card'
 require 'health_cards/chunking'
