+++ lib/health_cards/key.rb
@@ -17,11 +17,11 @@ module HealthCards
     end
 
     def to_jwk
-      required.merge(thumbprint: thumbprint, use: 'sig', alg: 'ES256')
+      required_attributes.merge(thumbprint: thumbprint, use: 'sig', alg: 'ES256')
     end
 
     def thumbprint
-      Base64.urlsafe_encode64(DIGEST.digest(required.except(:d).to_json), padding: false)
+      Base64.urlsafe_encode64(DIGEST.digest(required_attributes.except(:d).to_json), padding: false)
     end
 
     # Represents a set of Public Keys
@@ -37,7 +37,7 @@ module HealthCards
       end
     end
 
-    def required
+    def required_attributes
       BASE.merge(coordinates)
     end
 
@@ -50,51 +50,6 @@ module HealthCards
       @coordinates = coor.merge(xy).transform_values do |val|
         Base64.urlsafe_encode64([val].pack('H*'), padding: false)
       end
-    end
-  end
-
-  # A key used for singing JWS
-  class PrivateKey < Key
-    def self.from_file(path)
-      pem = OpenSSL::PKey::EC.new(File.read(path))
-      PrivateKey.new(pem)
-    end
-
-    def self.from_file!(path)
-      if File.exist?(path)
-        from_file(path)
-      else
-        generate_key(file_path: path)
-      end
-    end
-
-    def self.generate_key(file_path: nil)
-      key = OpenSSL::PKey::EC.generate('prime256v1')
-      File.write(file_path, key.to_pem) if file_path
-      PrivateKey.new(key)
-    end
-
-    def sign(payload)
-      @key.dsa_sign_asn1(payload)
-    end
-
-    def public_key
-      return @public_key if @public_key
-
-      pub = OpenSSL::PKey::EC.new('prime256v1')
-      pub.public_key = @key.public_key
-      @public_key = PublicKey.new(pub)
-    end
-  end
-
-  # A key used for verifying JWS
-  class PublicKey < Key
-    def self.from_json(json)
-      # TODO
-    end
-
-    def verify(payload, signature)
-      @key.dsa_verify_asn1(payload, signature)
     end
   end
 end
